Added warning dialogs, Menu bar working,  update readme
Delete button warning. Error Dialog for directory & file type.

+++ src/main/MainController.java
@@ -4,18 +4,19 @@ package main;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.*;
+import javafx.scene.control.Button;
+import javafx.scene.control.MenuItem;
 import javafx.stage.DirectoryChooser;
-import javafx.stage.FileChooser;
-import javafx.stage.Stage;
+import org.controlsfx.control.action.Action;
+import org.controlsfx.dialog.*;
+import org.controlsfx.dialog.Dialog;
 
-import javax.swing.*;
-import java.io.File;
-import java.io.FilenameFilter;
+import java.awt.*;
+import java.io.*;
 import java.net.URL;
-import java.nio.file.Files;
 import java.util.ResourceBundle;
 
-public class MainController implements Initializable{
+public class MainController implements Initializable {
 
     @FXML
     private Button chooseDirectory;
@@ -30,6 +31,7 @@ public class MainController implements Initializable{
     private MenuItem oneDayID, oneWeekID, twoWeeksID, oneMonthID, twoMonthsID;
 
     File selectedDirectory;
+    Byte directoryChoosenCount = 0;
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
@@ -39,6 +41,7 @@ public class MainController implements Initializable{
     }
 
     public void openFolder() {
+
         DirectoryChooser chooser = new DirectoryChooser();
         chooser.setTitle("Choose a Folder");
 
@@ -51,8 +54,11 @@ public class MainController implements Initializable{
         //Gets folder name
         String[] temp = selectedDirectory.getPath().split("/");
 
+        directoryChoosenCount = 1;
+
         //Adds folder name to "Directory" button
-        chooseDirectory.setText("Directory Choosen: " + temp[temp.length-1]);
+        chooseDirectory.setText("Directory Choosen: " + temp[temp.length - 1]);
+
         System.out.println(selectedDirectory);
     }
 
@@ -75,6 +81,7 @@ public class MainController implements Initializable{
     public void oneMonth() {
         datePicker.getEditor().setText(oneMonthID.getText());
     }
+
     //Sets the text of the DatePicker box to twoMonths
     public void twoMonths() {
         datePicker.getEditor().setText(twoMonthsID.getText());
@@ -82,33 +89,60 @@ public class MainController implements Initializable{
 
     //Calls deleteFile() to delete selected file types
     public void deleteButton() {
-        System.out.println("hey");
-        System.out.println(comboBox.getEditor().getText());
         deleteFile();
     }
 
-    public void deleteFile () {
+    public void deleteFile() {
         final File folder = selectedDirectory;
+
+        if (directoryChoosenCount == 0) {
+            Dialogs.create()
+                    .owner(Main.stage)
+                    .title("Error Dialog")
+                    .masthead("Directory Not Choosen")
+                    .message("Choose directory by clicking on the 'Choose Directory' button.")
+                    .showError();
+        }
+
+        if (comboBox.getEditor().getText().equals("")) {
+            Dialogs.create()
+                    .owner(Main.stage)
+                    .title("Error Dialog")
+                    .masthead("File Type Not Chosen")
+                    .message("Choose file type by clicking on arrow attached to 'File Type' box.")
+                    .showError();
+        }
+
         //Gets files based on the file type selected and puts them into an array
-        final File[] files = folder.listFiles( new FilenameFilter() {
+        final File[] files = folder.listFiles(new FilenameFilter() {
             @Override
-            public boolean accept( final File dir, final String name ) {
-                return name.matches( ".*\\" + comboBox.getEditor().getText());
+            public boolean accept(final File dir, final String name) {
+                return name.matches(".*\\" + comboBox.getEditor().getText());
             }
-        } );
+        });
         //Goes through files[] and deletes the files one by one
-        for ( final File file : files ) {
-            if (datePicker.getEditor().getText().equals("")) {
-                if (!file.delete()) {
-                    System.err.println("Can't remove " + file.getAbsolutePath());
+        Action response = Dialogs.create()
+                .owner(Main.stage)
+                .title("Confirm Dialog")
+                .masthead("You cannot recover the files!!!")
+                .message("Do you want to continue?")
+                .showConfirm();
+
+        if (response == Dialog.ACTION_YES) {
+            System.out.println("yes");
+            for (final File file : files) {
+                if (datePicker.getEditor().getText().equals("")) {
+                    if (!file.delete()) {
+                        System.err.println("Can't remove " + file.getAbsolutePath());
+                    }
+                } else {
+                    lastModified(file);
                 }
-            } else {
-               lastModified(file);
             }
         }
     }
 
-    public void lastModified (File timedFile) {
+    public void lastModified(File timedFile) {
         //Deletes files if last modified seconds is lest than a day
         if (datePicker.getEditor().getText().equals("One Day")) {
             if ((System.currentTimeMillis() - timedFile.lastModified()) <= 86400000) {
@@ -149,5 +183,15 @@ public class MainController implements Initializable{
                 }
             }
         }
+    }
+
+    public void openAbout() throws IOException {
+        //Opens text file with notepad
+        File about = new File("src/about.txt");
+        Desktop.getDesktop().edit(about);
+    }
+
+    public void close() {
+        Main.stage.close();
     }
 }
